+++ pylinkage/interface/__init__.py
@@ -3,12 +3,13 @@ Classes that represent the main interface with the library.
 """
 from .exceptions import (
     UnbuildableError,
-    HypostaticError
+    HypostaticError,
+    NotCompletelyDefinedError,
 )
 from .joint import (
     Static,
     Crank,
-    Fixed
+    Fixed,
 )
-from .revolute_joint import Revolute, Pivot
+from ..joints.revolute import Revolute, Pivot
 from .linkage import Linkage