+++ pylinkage/interface/exceptions.py
@@ -31,8 +31,15 @@ class UnbuildableError(Exception):
 
 
 class HypostaticError(Exception):
-    """The system is under-constrained and multiple solutions may exist."""
+    """The linkage is under-constrained and multiple solutions may exist."""
 
-    def __init__(self, linkage, message='The system is hypo-static!'):
+    def __init__(self, linkage, message='The linkage is under-constrained!'):
         self.linkage = linkage
+        super().__init__(message)
+
+
+class NotCompletelyDefinedError(Exception):
+    """The linkage definition is incomplete."""
+    def __init__(self, joint, message='The joint is not completely defined!'):
+        self.joint = joint
         super().__init__(message)