+++ pylinkage/joints/__init__.py
@@ -1,0 +1,6 @@
+"""Definition of joints."""
+
+from .linear import Linear
+from .revolute import Revolute
+
+from ..interface.joint import (Static, Fixed, Crank,)