+++ tests/test_joints.py
@@ -8,50 +8,51 @@ Created on Fri Apr 16 17:19:47 2021
 
 import unittest
 import math
+
 from pylinkage import UnbuildableError
 from pylinkage.interface.joint import Fixed
-from pylinkage.interface.revolute_joint import Revolute, Pivot
+from pylinkage.joints import Revolute
 
 
 class TestPivot(unittest.TestCase):
     """Test Pivot Joint."""
 
-    pivot1 = Pivot(0, 0)
+    pivot1 = Revolute(0, 0)
 
     def test_buildable(self):
         """Upper intersect test."""
-        pivot2 = Pivot(1, 0)
-        pivot3 = Pivot(
+        pivot2 = Revolute(1, 0)
+        pivot3 = Revolute(
             y=1, joint0=self.pivot1, joint1=pivot2,
             distance0=1, distance1=1
         )
         pivot3.reload()
-        self.assertEqual(pivot3.coord(), (.5, math.sqrt(.75)))
+        self.assertEqual((.5, math.sqrt(.75)), pivot3.coord())
 
     def test_under_intersect(self):
         """Under intersect test."""
-        pivot2 = Pivot(1, 0)
-        pivot3 = Pivot(
+        pivot2 = Revolute(1, 0)
+        pivot3 = Revolute(
             y=-1, joint0=self.pivot1, joint1=pivot2,
             distance0=1, distance1=1
         )
         pivot3.reload()
-        self.assertEqual(pivot3.coord(), (.5, -math.sqrt(.75)))
+        self.assertEqual((.5, -math.sqrt(.75)), pivot3.coord())
 
     def test_limit_intersect(self):
         """Test system almost breaking."""
-        pivot2 = Pivot(2, 0)
-        pivot3 = Pivot(
+        pivot2 = Revolute(2, 0)
+        pivot3 = Revolute(
             y=1, joint0=self.pivot1, joint1=pivot2,
             distance0=1, distance1=1
         )
         pivot3.reload()
-        self.assertEqual(pivot3.coord(), (1, 0))
+        self.assertEqual((1, 0), pivot3.coord())
 
     def test_no_intersect(self):
         """Test system almost breaking."""
-        pivot2 = Pivot(0, 3)
-        pivot3 = Pivot(
+        pivot2 = Revolute(0, 3)
+        pivot3 = Revolute(
             y=1, joint0=self.pivot1, joint1=pivot2,
             distance0=1, distance1=1
         )
@@ -72,7 +73,7 @@ class TestRevolute(unittest.TestCase):
             distance0=1, distance1=1
         )
         pivot3.reload()
-        self.assertEqual(pivot3.coord(), (.5, math.sqrt(.75)))
+        self.assertEqual((.5, math.sqrt(.75)), pivot3.coord())
 
     def test_under_intersect(self):
         """Under intersect test."""
@@ -82,7 +83,7 @@ class TestRevolute(unittest.TestCase):
             distance0=1, distance1=1
         )
         pivot3.reload()
-        self.assertEqual(pivot3.coord(), (.5, -math.sqrt(.75)))
+        self.assertEqual((.5, -math.sqrt(.75)), pivot3.coord())
 
     def test_limit_intersect(self):
         """Test system almost breaking."""
@@ -92,7 +93,7 @@ class TestRevolute(unittest.TestCase):
             distance0=1, distance1=1
         )
         pivot3.reload()
-        self.assertEqual(pivot3.coord(), (1, 0))
+        self.assertEqual((1, 0), pivot3.coord())
 
     def test_no_intersect(self):
         """Test system almost breaking."""
@@ -108,16 +109,16 @@ class TestRevolute(unittest.TestCase):
 class TestFixed(unittest.TestCase):
     """Test Fixed_Joint."""
 
-    pivot1 = Pivot(0, 0)
+    pivot1 = Revolute(0, 0)
 
     def test_pos(self):
         """Test Fixed_Joint positioning."""
-        pivot2 = Pivot(1, 0)
+        pivot2 = Revolute(1, 0)
         fixed = Fixed(
             joint0=self.pivot1, joint1=pivot2, angle=0, distance=1
         )
         fixed.reload()
-        self.assertEqual(fixed.coord(), (1, 0))
+        self.assertEqual((1, 0), fixed.coord())
 
 
 if __name__ == '__main__':